Made a pure virtual function and changed BankAccount to an abstract class

+++ src/examples/04_module/01_bank/checking_account.h
@@ -6,5 +6,5 @@ class CheckingAccount : public BankAccount
 public:
 	CheckingAccount() = default;
 	explicit CheckingAccount(int b) : BankAccount(b) {}
-	double get_balance()const { return balance * (1 + get_rate()); }
+	int get_balance()const { return balance * (1 + get_rate()); }
 };
